requirements.txt
aiogram
python-dotenv
aioboto3

